pyproject.toml
[project]
name = "practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dds>=1.0",
    "glfw>=2.10.0",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pygame>=2.6.1",
    "pyglm>=2.8.2",
    "pyopengl>=3.1.10",
    "pyrr>=0.10.3",
]
